# ServerAPI/src/api/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"FriendRequestDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"senderId": {
						"type": "string"
					},
					"receiverId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"senderId",
					"receiverId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					},
					"twitter": {
						"type": "string"
					},
					"linkedin": {
						"type": "string"
					},
					"facebook": {
						"type": "string"
					},
					"instagram": {
						"type": "string"
					},
					"telegram": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserPersonalDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					},
					"twitter": {
						"type": "string"
					},
					"linkedin": {
						"type": "string"
					},
					"facebook": {
						"type": "string"
					},
					"instagram": {
						"type": "string"
					},
					"telegram": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"SentRequests": {
						"items": {
							"$ref": "#/components/schemas/FriendRequestDTO"
						},
						"type": "array"
					},
					"ReceivedRequests": {
						"items": {
							"$ref": "#/components/schemas/FriendRequestDTO"
						},
						"type": "array"
					},
					"Friends": {
						"items": {
							"$ref": "#/components/schemas/UserDTO"
						},
						"type": "array"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateLinksDTO": {
				"properties": {
					"twitter": {
						"type": "string"
					},
					"linkedin": {
						"type": "string"
					},
					"facebook": {
						"type": "string"
					},
					"instagram": {
						"type": "string"
					},
					"telegram": {
						"type": "string"
					},
					"website": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUsernameDTO": {
				"properties": {
					"username": {
						"type": "string"
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateFullNameDTO": {
				"properties": {
					"fullName": {
						"type": "string"
					}
				},
				"required": [
					"fullName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseSuccessType_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"response": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"error",
					"response"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FriendRequestCreateDTO": {
				"properties": {
					"receiverId": {
						"type": "string"
					}
				},
				"required": [
					"receiverId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FriendRequestDetailDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"senderId": {
						"type": "string"
					},
					"receiverId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"Sender": {
						"$ref": "#/components/schemas/UserDTO"
					},
					"Receiver": {
						"$ref": "#/components/schemas/UserDTO"
					}
				},
				"required": [
					"id",
					"senderId",
					"receiverId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseSuccessType_FriendRequestDetailDTO_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"response": {
						"$ref": "#/components/schemas/FriendRequestDetailDTO"
					}
				},
				"required": [
					"success",
					"error",
					"response"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseSuccessType_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"response": {}
				},
				"required": [
					"success",
					"error",
					"response"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FriendMinimalDTO": {
				"properties": {
					"friends": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"friends"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseSuccessType_FriendMinimalDTO_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"response": {
						"$ref": "#/components/schemas/FriendMinimalDTO"
					}
				},
				"required": [
					"success",
					"error",
					"response"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthTokenDTO": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseSuccessType_AuthTokenDTO_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"response": {
						"$ref": "#/components/schemas/AuthTokenDTO"
					}
				},
				"required": [
					"success",
					"error",
					"response"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthDTO": {
				"properties": {
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserCreateDTO": {
				"properties": {
					"username": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"fullName"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "serverapi",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/profile": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserPersonalDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/profile/update_links": {
			"put": {
				"operationId": "UpdateLinks",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateLinksDTO"
							}
						}
					}
				}
			}
		},
		"/profile/update_username": {
			"put": {
				"operationId": "UpdateUsername",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUsernameDTO"
							}
						}
					}
				}
			}
		},
		"/profile/update_fullName": {
			"put": {
				"operationId": "UpdateFullName",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Profile"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateFullNameDTO"
							}
						}
					}
				}
			}
		},
		"/friend_request": {
			"post": {
				"operationId": "CreateFriendRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseSuccessType_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Friend Request"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FriendRequestCreateDTO"
							}
						}
					}
				}
			}
		},
		"/friend_request/{id}": {
			"put": {
				"operationId": "AcceptFriendRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseSuccessType_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Friend Request"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "RejectFriendRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseSuccessType_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Friend Request"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/friend_request/sent": {
			"get": {
				"operationId": "GetSentFriendRequests",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseSuccessType_FriendRequestDetailDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"Friend Request"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/friend_request/received": {
			"get": {
				"operationId": "GetFriendRequests",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseSuccessType_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Friend Request"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/{id}/friends": {
			"get": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseSuccessType_FriendMinimalDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"Friends"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseSuccessType_AuthTokenDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/AuthDTO"
									},
									{
										"$ref": "#/components/schemas/UserCreateDTO"
									}
								]
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseSuccessType_AuthTokenDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthDTO"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}